initial refactor
This moves things around and sets up a pattern where
DocumentContentModel can be compose of multiple types
that extend the BaseDocumentContentModel
